% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_map_crds.R
\name{iso_map_crds}
\alias{iso_map_crds}
\title{Map a list of sample_ids to CRDS data}
\usage{
iso_map_crds(crds_data, sample_map)
}
\arguments{
\item{crds_data}{cavity ring-down spectroscopy with a datetime column. Ideally this has been read using 'iso_read_crds'}

\item{sample_map}{a three-column tibble containing sample_id, datetime_start, and datetime_end, corresponding to sample intervals}
}
\value{
a tibble of CRDS data with a sample_id column mapped to specific time intervals
}
\description{
Map a list of sample IDs to CRDS data. Import a sample map with a list of sample_ids and two columns that correspond to the start and end datetimes of the sample
}
