% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_plot_crds.R
\name{iso_plot_crds}
\alias{iso_plot_crds}
\title{Plot CRDS data in a stackplot}
\usage{
iso_plot_crds(crds_data)
}
\arguments{
\item{crds_data}{a crds tibble}

\item{panels}{a character vector of the columns to be plotted}
}
\description{
Create a multi-plot of CRDS data for review.
}
