% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_read_crds.R
\name{iso_read_crds}
\alias{iso_read_crds}
\title{read Picarro CRDS output .dat file}
\usage{
iso_read_crds(path)
}
\arguments{
\item{path}{file path to a single CRDS .dat file}
}
\value{
a tibble of cleaned Picarro data
}
\description{
reads Picarro cavity ring down spectroscopy 'CRDS' .dat file
}
\examples{
# read a data file
iso_read_crds("a_crds_data_file.dat")
# pass the data to a variable in your R workspace
df <- iso_read_crds("a_crds_data_file.dat")
df
}
