% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_read_many_crds.R
\name{iso_read_many_crds}
\alias{iso_read_many_crds}
\title{read multiple Picarro CRDS output .dat file}
\usage{
iso_read_many_crds(directory)
}
\arguments{
\item{directory}{a folder directory containing multiple .dat files}
}
\value{
a tibble of cleaned Picarro data
}
\description{
reads all Picarro cavity ring down spectroscopy 'CRDS' .dat files from a single directory.
}
